rename some files & remove redux from navbar components
