+++ src/components/Navbar/Watchlists/CreateWatchlistBtn.js
@@ -2,7 +2,7 @@ import React from 'react'
 import cx from 'classnames'
 import Button from '@santiment-network/ui/Button'
 import NewWatchlist from '../../../ducks/Watchlists/Actions/New'
-import styles from './WatchlistsEmptySection.module.scss'
+import styles from './EmptySection.module.scss'
 
 const CreateWatchlistBtn = ({ className, watchlists }) => (
   <NewWatchlist
