+++ src/components/Navbar/Watchlists/EmptySection.js
@@ -3,7 +3,7 @@ import Button from '@santiment-network/ui/Button'
 import { SvgNew } from '../../Watchlists/NewWatchlistCard'
 import NewWatchlist from '../../../ducks/Watchlists/Actions/New'
 import CreateWatchlistBtn from './CreateWatchlistBtn'
-import styles from './WatchlistsEmptySection.module.scss'
+import styles from './EmptySection.module.scss'
 
 const WatchlistsEmptySection = ({ watchlists }) => (
   <div className={styles.wrapper}>
