+++ src/components/Navbar/Watchlists/EmptySection.module.scss
