+++ src/components/Navbar/Watchlists/MarketDropdown.module.scss
@@ -1,5 +1,5 @@
-@import '../../../../node_modules/@santiment-network/ui/variables';
-@import '../../../../node_modules/@santiment-network/ui/mixins';
+@import '~@santiment-network/ui/variables';
+@import '~@santiment-network/ui/mixins';
 
 .wrapper {
   display: flex;
