+++ src/components/Navbar/Watchlists/WatchlistsDropdown.module.scss
@@ -1,5 +1,4 @@
-@import '../../../../node_modules/@santiment-network/ui/variables';
-@import '../../../../node_modules/@santiment-network/ui/mixins';
+@import '~@santiment-network/ui/mixins';
 
 .wrapper {
   box-sizing: border-box;
